changed Double-BigDecimal for better money handling

+++ src/main/java/com/example/balancee_rewards/dto/CustomerDTO.java
@@ -1,14 +1,16 @@
 package com.example.balancee_rewards.dto;
 
+import java.math.BigDecimal;
+
 public class CustomerDTO {
     private String transactionId;
-    private Double totalCashback;
-    private Double currentBalance;
+    private BigDecimal totalCashback;
+    private BigDecimal currentBalance;
 
     public CustomerDTO(){
     }
 
-    public CustomerDTO(String transactionId, Double totalCashback, Double currentBalance) {
+    public CustomerDTO(String transactionId, BigDecimal totalCashback, BigDecimal currentBalance) {
         this.transactionId = transactionId;
         this.totalCashback = totalCashback;
         this.currentBalance = currentBalance;
@@ -18,11 +20,11 @@ public class CustomerDTO {
         return transactionId;
     }
 
-    public Double getCurrentBalance() {
+    public BigDecimal getCurrentBalance() {
         return currentBalance;
     }
 
-    public Double getTotalCashback() {
+    public BigDecimal getTotalCashback() {
         return totalCashback;
     }
 
@@ -30,11 +32,11 @@ public class CustomerDTO {
         this.transactionId = transactionId;
     }
 
-    public void setCurrentBalance(Double currentBalance) {
+    public void setCurrentBalance(BigDecimal currentBalance) {
         this.currentBalance = currentBalance;
     }
 
-    public void setTotalCashback(Double totalCashback) {
+    public void setTotalCashback(BigDecimal totalCashback) {
         this.totalCashback = totalCashback;
     }
 }
